backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
requests
python-dotenv
pydantic
aiohttp
beautifulsoup4
firebase-admin
python-multipart
